+++ src/app/core/models/app-settings.ts
@@ -1,5 +1,6 @@
 export interface AppSettings {
   color?: ColorType;
+  addresses?: string[];
 }
 
 export type ColorType = 'default' | 'white' | 'blue';