+++ src/app/orders/orders.module.ts
@@ -1,10 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 
 import { OrderFormComponent } from './order-form.component';
 import { OrderService } from './services/order.service';
+import { ValidatorsModule } from '../validators/validators.module';
 
 @NgModule({
   declarations: [
@@ -13,7 +14,9 @@ import { OrderService } from './services/order.service';
   imports: [
     CommonModule,
     FormsModule,
-    RouterModule
+    RouterModule,
+    ReactiveFormsModule,
+    ValidatorsModule
   ],
   providers: [
     OrderService